add support for multiple command words

+++ src/main/java/seedu/doist/logic/parser/Parser.java
@@ -42,38 +42,58 @@ public class Parser {
 
         final String commandWord = matcher.group("commandWord");
         final String arguments = matcher.group("arguments");
-        switch (commandWord) {
-
-        case AddCommand.COMMAND_WORD:
-            return new AddCommandParser().parse(arguments);
-
-        case EditCommand.COMMAND_WORD:
-            return new EditCommandParser().parse(arguments);
-
-        case SelectCommand.COMMAND_WORD:
-            return new SelectCommandParser().parse(arguments);
-
-        case DeleteCommand.COMMAND_WORD:
-            return new DeleteCommandParser().parse(arguments);
-
-        case ClearCommand.COMMAND_WORD:
-            return new ClearCommand();
-
-        case FindCommand.COMMAND_WORD:
-            return new FindCommandParser().parse(arguments);
-
-        case ListCommand.COMMAND_WORD:
-            return new ListCommand();
-
-        case ExitCommand.COMMAND_WORD:
-            return new ExitCommand();
-
-        case HelpCommand.COMMAND_WORD:
-            return new HelpCommand();
-
-        default:
-            return new IncorrectCommand(MESSAGE_UNKNOWN_COMMAND);
+        if (AddCommand.commandWords.contains(commandWord) || AddCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new AddCommandParser().parse(arguments);
+        } else if (EditCommand.commandWords.contains(commandWord) || EditCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new EditCommandParser().parse(arguments);
+        } else if (SelectCommand.commandWords.contains(commandWord) || SelectCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new SelectCommandParser().parse(arguments);
+        } else if (DeleteCommand.commandWords.contains(commandWord) || DeleteCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new DeleteCommandParser().parse(arguments);
+        } else if (ClearCommand.commandWords.contains(commandWord) || ClearCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new ClearCommand();
+        } else if (FindCommand.commandWords.contains(commandWord) || FindCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new FindCommandParser().parse(arguments);
+        } else if (ListCommand.commandWords.contains(commandWord) || ListCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new ListCommand();
+        } else if (ExitCommand.commandWords.contains(commandWord) || ExitCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new ExitCommand();
+        } else if (HelpCommand.commandWords.contains(commandWord) || HelpCommand.DEFAULT_COMMAND_WORD == commandWord) {
+        	return new HelpCommand();
+        } else {
+        	return new IncorrectCommand(MESSAGE_UNKNOWN_COMMAND);
         }
+//        switch (commandWord) {
+//
+//        case AddCommand.COMMAND_WORD:
+//            return new AddCommandParser().parse(arguments);
+//
+//        case EditCommand.COMMAND_WORD:
+//            return new EditCommandParser().parse(arguments);
+//
+//        case SelectCommand.COMMAND_WORD:
+//            return new SelectCommandParser().parse(arguments);
+//
+//        case DeleteCommand.COMMAND_WORD:
+//            return new DeleteCommandParser().parse(arguments);
+//
+//        case ClearCommand.COMMAND_WORD:
+//            return new ClearCommand();
+//
+//        case FindCommand.COMMAND_WORD:
+//            return new FindCommandParser().parse(arguments);
+//
+//        case ListCommand.COMMAND_WORD:
+//            return new ListCommand();
+//
+//        case ExitCommand.COMMAND_WORD:
+//            return new ExitCommand();
+//
+//        case HelpCommand.COMMAND_WORD:
+//            return new HelpCommand();
+//
+//        default:
+//            return new IncorrectCommand(MESSAGE_UNKNOWN_COMMAND);
+//        }
     }
-
 }
